Fix missing query param

+++ src/main/java/de/chojo/repbot/dao/access/user/sub/UserToken.java
@@ -11,6 +11,7 @@ import de.chojo.sadu.mapper.wrapper.Row;
 import java.sql.SQLException;
 import java.time.Instant;
 
+import static de.chojo.sadu.queries.api.call.Call.call;
 import static de.chojo.sadu.queries.api.query.Query.query;
 import static de.chojo.sadu.queries.converter.StandardValueConverter.INSTANT_TIMESTAMP;
 
@@ -25,6 +26,6 @@ public record UserToken(long userId, String accessToken, String refreshToken, In
     }
 
     public void delete() {
-        query("DELETE FROM user_token WHERE user_id = ?").single().delete();
+        query("DELETE FROM user_token WHERE user_id = ?").single(call().bind(userId)).delete();
     }
 }
